1.Verified using of printHello method in GreetingControllerTests 2. Renamed "controller" package to "controllers" 3. Removed commented line in Main class
