+++ src/main/java/com/example/demo/Application.java
@@ -6,7 +6,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) {
-        //ApplicationContext context = new AnnotationConfigApplicationContext("com.example.demo.services");
         SpringApplication.run(Application.class, args);
     }
 }
